Style: ゲーム取得時のゲームの表示を更新
