+++ src/features/acquisition/components/SearchResults.tsx
@@ -1,4 +1,4 @@
-import { Button, Center, Image, Skeleton } from "@mantine/core";
+import { Button, Center, Image, SimpleGrid, Skeleton } from "@mantine/core";
 import { itemsToShowAtom } from "atoms/games";
 import {
   gameResultsAtom,
@@ -8,7 +8,7 @@ import { useAtom } from "jotai";
 import { useMediaQuery } from "lib/mantine/useMediaQuery";
 
 import before_search from "../../../assets/before_search.png";
-import GameCardAcquisition from "./GameCardAcquisition";
+import Game from "./Game";
 
 const SearchResults = () => {
   const [games] = useAtom(gameResultsAtom);
@@ -42,19 +42,21 @@ const SearchResults = () => {
   }
 
   return (
-    <div className="flex flex-wrap justify-between gap-4">
-      {games.slice(0, itemsToShow).map((game) => (
-        <GameCardAcquisition key={game.title} game={game} />
-      ))}
+    <>
+      <SimpleGrid cols={largerThanSm ? 2 : 1}>
+        {games.slice(0, itemsToShow).map((game) => (
+          <Game key={game.url} game={game} />
+        ))}
+      </SimpleGrid>
       <Button
         onClick={() => setItemsToShow(itemsToShow + 30)}
-        className="flex w-full items-center justify-center border-0 border-y border-gray-300 bg-white text-black hover:bg-gray-100"
+        className="mt-4 flex w-full items-center justify-center border-0 border-y border-gray-300 bg-white text-black hover:bg-gray-100"
         size="md"
         style={{ display: itemsToShow >= games.length ? "none" : "flex" }}
       >
         さらに表示する
       </Button>
-    </div>
+    </>
   );
 };
